Added solving linear equation.

+++ regression/src/DataGenerator.cpp
@@ -9,8 +9,8 @@ void DataGenerator::generateDataSin(unsigned int& number, vector<DataPoint>& dat
 {
 	for( int i = 0; i < number; ++i )
 	{
-		float x = (float)( i / 10.0f );
-		float result = sin( (2.0 * M_PI * x) );
+		double x = (double)( i / 10.0 );
+		double result = sin( (2.0 * M_PI * x) );
 		DataPoint p(i, result);
 		dataPoints.push_back(p);
 	}
@@ -20,12 +20,12 @@ void DataGenerator::generateDataSinNoise(unsigned int& number, vector<DataPoint>
 {
 	for( int i = 0; i < number; ++i )
 	{
-		float x = (float)( i / 10.0f );
+		double x = (double)( i / 10.0 );
 
-		float result = sin( (2.0 * M_PI * x) );
+		double result = sin( (2.0 * M_PI * x) );
 		
-		float noise_x = (float)((rand() % 100) / 500.0f);
-		float noise_y = (float)((rand() % 100) / 500.0f);
+		double noise_x = (double)((rand() % 100) / 500.0f);
+		double noise_y = (double)((rand() % 100) / 500.0f);
 
 		DataPoint p(i + (noise_x+noise_y), result + (noise_y-noise_x));
 		dataPoints.push_back(p);
